Add an automatically transcribed dict addendum
Added a transcr_dict in addition to the default prondict and custom dict to collect automatically transcribed words.
This should speed up batch transcription greatly as each word only gets transcribed once by the LSTM, and is cached after that.
Also opens up the option to dump the automatically transcribed dictionary to a file afterwards, e.g. to eliminate OOV in a given corpus.